grid/pyproject.toml
[tool.poetry]
name = "svg_browser"
version = "0.1.0"
description = "Lokalny serwer do przeglądania plików SVG z kategoryzacją PWA i metadanych"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
flask = "^2.3.0"
python-dotenv = "^1.0.0"
lxml = "^4.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
svg_browser = "svg_browser.app:main"